chore: restructure AGP test coverage — existing apps to AGP 9, minimal AGP 8 smoke
Both flutter_sample_spm and flutter_sample_cocoapods now build under
AGP 9.0.1 + Gradle 9.1.0 with android.builtInKotlin=true, exercising
the AGP 9 / built-in Kotlin path with realistic plugin loads (Firebase,
local_notifications, etc).

flutter_sample_agp9_test is replaced by flutter_sample_agp8_min — a
stripped-down copy of spm that keeps AGP 8.9.3 + KGP 2.1.21. It is the
smoke target ensuring the plugin's dynamic AGP-version check still
applies kotlin-android correctly on AGP < 9.

build-agp9-test.yml is replaced by build-agp8-min-test.yml accordingly;
the existing build-sample-apps.yml workflow now covers the AGP 9 path.

Risk: upstream plugins consumed by the two real sample apps
(flutter_local_notifications, firebase_messaging, etc.) may not yet
support AGP 9 built-in Kotlin. If CI surfaces failures inside their
plugin code rather than ours, flip android.builtInKotlin=false to keep
AGP 9 with legacy KGP application as the fallback.
