+++ apps/flutter_sample_agp8_min/android/app/build.gradle
@@ -1,10 +1,11 @@
 plugins {
     id "com.android.application"
+    id "kotlin-android"
     id "dev.flutter.flutter-gradle-plugin"
 }
 
 android {
-    namespace 'io.customer.testbed.flutter.agp9test'
+    namespace 'io.customer.testbed.flutter.agp8min'
     compileSdkVersion 36
     ndkVersion flutter.ndkVersion
 
@@ -13,18 +14,16 @@ android {
         targetCompatibility JavaVersion.VERSION_17
     }
 
+    kotlinOptions {
+        jvmTarget = '17'
+    }
+
     defaultConfig {
-        applicationId "io.customer.testbed.flutter.agp9test"
-        minSdkVersion 21
+        applicationId "io.customer.testbed.flutter.agp8min"
+        minSdkVersion flutter.minSdkVersion
         targetSdkVersion 36
         versionCode 1
         versionName "1.0"
-    }
-}
-
-kotlin {
-    compilerOptions {
-        jvmTarget = org.jetbrains.kotlin.gradle.dsl.JvmTarget.JVM_17
     }
 }
