+++ apps/flutter_sample_agp8_min/android/app/src/main/kotlin/io/customer/testbed/flutter/agp8min/MainActivity.kt
@@ -1,4 +1,4 @@
-package io.customer.testbed.flutter.agp9test
+package io.customer.testbed.flutter.agp8min
 
 import io.flutter.embedding.android.FlutterActivity
