+++ apps/flutter_sample_agp8_min/android/build.gradle
