+++ apps/flutter_sample_agp8_min/android/gradle.properties
@@ -1,5 +1,3 @@
 org.gradle.jvmargs=-Xmx4g
 android.useAndroidX=true
 android.enableJetifier=true
-android.builtInKotlin=true
-customerio_location_enabled=true
