+++ apps/flutter_sample_agp8_min/android/settings.gradle
@@ -18,7 +18,8 @@ pluginManagement {
 
 plugins {
     id "dev.flutter.flutter-plugin-loader" version "1.0.0"
-    id "com.android.application" version "9.0.1" apply false
+    id "com.android.application" version "8.9.3" apply false
+    id "org.jetbrains.kotlin.android" version "2.1.21" apply false
 }
 
 include ':app'
