--- apps/flutter_sample_agp9_test/.gitignore
@@ -1,21 +1,0 @@
-# Flutter / Dart
-.dart_tool/
-.flutter-plugins
-.flutter-plugins-dependencies
-.packages
-.pub-cache/
-.pub/
-build/
-pubspec.lock
-
-# Android
-android/.gradle/
-android/local.properties
-android/app/debug.keystore
-android/captures/
-android/key.properties
-*.iml
-
-# IDE
-.idea/
-.vscode/
