--- apps/flutter_sample_agp9_test/android/gradle.properties
@@ -1,6 +1,0 @@
-org.gradle.jvmargs=-Xmx4g
-android.useAndroidX=true
-android.enableJetifier=true
-# Opt into AGP 9.0's built-in Kotlin support. Flutter defaults this to false
-# to preserve legacy behaviour; this app verifies the opposite path.
-android.builtInKotlin=true
