+++ apps/flutter_sample_spm/android/gradle.properties
@@ -1,4 +1,5 @@
 org.gradle.jvmargs=-Xmx4g
 android.useAndroidX=true
 android.enableJetifier=true
+android.builtInKotlin=true
 customerio_location_enabled=true
