+++ apps/flutter_sample_spm/android/settings.gradle
@@ -18,8 +18,7 @@ pluginManagement {
 
 plugins {
     id "dev.flutter.flutter-plugin-loader" version "1.0.0"
-    id "com.android.application" version "8.9.3" apply false
-    id "org.jetbrains.kotlin.android" version "2.1.21" apply false
+    id "com.android.application" version "9.0.1" apply false
     id "com.google.gms.google-services" version "4.3.15" apply false
 }
